+++ modules/net/include/dap_chain_ledger.h
@@ -301,8 +301,8 @@ json_object *dap_ledger_token_info_by_name(dap_ledger_t *a_ledger, const char *a
 // Get all token-declarations
 dap_list_t* dap_ledger_token_decl_all(dap_ledger_t *a_ledger);
 
-json_object *dap_ledger_threshold_info(dap_ledger_t *a_ledger, size_t a_limit, size_t a_offset, dap_hash_fast_t *a_threshold_hash);
-json_object *dap_ledger_balance_info(dap_ledger_t *a_ledger, size_t a_limit, size_t a_offset);
+json_object *dap_ledger_threshold_info(dap_ledger_t *a_ledger, int a_limit, size_t a_offset, dap_hash_fast_t *a_threshold_hash);
+json_object *dap_ledger_balance_info(dap_ledger_t *a_ledger, int a_limit, size_t a_offset);
 
 size_t dap_ledger_token_get_auth_signs_valid(dap_ledger_t *a_ledger, const char *a_token_ticker);
 size_t dap_ledger_token_get_auth_signs_total(dap_ledger_t *a_ledger, const char *a_token_ticker);
